Tests/LibWeb/Text/expected/HTML/FormAssociatedElement-set-selectionDirection.txt

select event fired for input
select event fired for textarea
